src/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
google-auth==2.17.3
